refactor: move hocs to components folder
