+++ src/components/hocs/with-figure-caption/Figure.module.scss
