+++ src/components/hocs/with-figure-caption/__tests__/with-figure-caption.test.tsx
