+++ src/components/hocs/with-figure-caption/with-figure-caption.tsx
