+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from './components/atoms';
+export * from './components/hocs';
 export * from './components/molecules';
 export * from './components/organisms';
 export * from './contexts';
