+++ vite.config.ts
@@ -53,7 +53,6 @@ export default defineConfig({
       { find: '@models', replacement: path.resolve(__dirname, 'src/models') },
       { find: '@contexts', replacement: path.resolve(__dirname, 'src/contexts') },
       { find: '@icons', replacement: path.resolve(__dirname, 'src/icons') },
-      { find: '@hocs', replacement: path.resolve(__dirname, 'src/hocs') },
       { find: '@styles', replacement: path.resolve(__dirname, 'src/styles') },
       { find: '@assets', replacement: path.resolve(__dirname, 'src/assets') },
       { find: '@components', replacement: path.resolve(__dirname, 'src/components') },
